SpaceShip work

+++ BJChat/src/spaceships/Bullet.java
@@ -5,14 +5,14 @@ import java.awt.Graphics;
 
 public class Bullet extends Actor  {
 	
-	public Bullet(int x, int y) {
+	public Bullet(double x, double y) {
 		super(x, y);
 	}
 	
 
 	public void draw(Graphics g){
 		g.setColor(Color.red);
-		g.fillOval(x, y, 10, 10);
+		g.fillOval((int)x, (int)y, 10, 10);
 	}
 
 }
